测量attention内存消耗
